+++ R/wflow_html.R
@@ -263,18 +263,10 @@ cache_hook <- function(x, options) {
 # Access the default chunk hook function from knitr because it isn't exported.
 knitr_hook_chunk <- function() {
 
-  input <- c("```{r get-chunk-hook, comment=NA, echo=FALSE}",
-             "body(knitr::knit_hooks$get('chunk'))",
-             "```")
-  outfile <- tempfile(fileext = ".md")
-  on.exit(fs::file_delete(outfile))
-  knitr::knit(text = input, output = outfile, quiet = TRUE, envir = new.env())
-  output <- readLines(outfile)
-  output <- stringr::str_subset(output, "```", negate = TRUE)
-  f <- function(x, options) {}
-  body(f) <- parse(text = output)
-  fence_char <- "`"
-  fence <- "```"
+  knitr::render_markdown()
+  f <- knitr::knit_hooks$get("chunk")
+  knitr::knit_hooks$restore()
+
   return(f)
 }
 
@@ -285,13 +277,6 @@ get_cache_hook <- function() {
   wflow_hook_chunk <- cache_hook
 
   result <- function(x, options) {
-
-    # Cannot currently use hidden knitr option `indent`, which is mainly useful
-    # for markdown output anways
-    if (!is.null(options$indent))
-      stop("The hidden knitr chunk option `indent` cannot be used with workflowr.\n",
-           glue::glue("It was set for the chunk labeled \"{options$label}\" in {knitr::current_input()}"),
-           call. = FALSE)
 
     x <- default_hook_chunk(x, options)
     wflow_hook_chunk(x, options)